refactor(desktop): declare the agent action shape once
